requirements.txt
# Core LangChain & LangGraph (0.3.x compatible series)
langchain>=0.3.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.21
langsmith>=0.1.147

# LLM Providers (compatible with 0.3.x)
langchain-cerebras>=0.3.0
langchain-google-genai>=1.0.10
google-generativeai>=0.7.0

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Utilities  
pydantic>=2.7.0
PyYAML>=6.0.1
requests>=2.32.3
pytest>=8.2.0
typing_extensions>=4.8.0
fastapi 
uvicorn[standard]
python-dotenv
proto-plus>=1.26.0
google-api-core>=2.15.0
langgraph-cli[inmem]